Separated ThreadSafeCounter into ThreadSafeCounterManager for Android and ThreadSafeCounterSharedMemory for everyone else. This was done to avoid random BrokenPipeError on system which support shared memory.
